Move client
