+++ strawberry/django/test/__init__.py
@@ -1,0 +1,4 @@
+from .client import GraphQLTestClient
+
+
+__all__ = ["GraphQLTestClient"]
