+++ strawberry/test/__init__.py
@@ -1,0 +1,4 @@
+from .client import BaseGraphQLTestClient, Body, Response
+
+
+__all__ = ["Body", "Response", "BaseGraphQLTestClient"]
