+++ tests/django/app/schema.py
@@ -1,19 +1,13 @@
 import strawberry
-from strawberry.types.info import Info
 
 
-def function_resolver(root) -> str:
-    return "I'm a function resolver"
+def hello_resolver(root) -> str:
+    return "🍓"
 
 
 @strawberry.type
 class Query:
-    @strawberry.field
-    def user(self, info: Info) -> str:
-        return "🍓"
-
-    hello: str = "🍓"
-    hello_field: str = strawberry.field(resolver=function_resolver)
+    hello: str = strawberry.field(resolver=hello_resolver)
 
 
 schema = strawberry.Schema(query=Query)
