+++ tests/django/fixtures.py
@@ -2,7 +2,7 @@ import pytest
 
 from django.test.client import Client
 
-from strawberry.django.test import GraphQLTestClient
+from strawberry.django.client import GraphQLTestClient
 
 
 @pytest.fixture()
