+++ tests/django/test_graphql_test_client.py
@@ -1,16 +1,5 @@
 def test_query(graphql_client):
     query = """query {
-        user
-    }
-    """
-
-    result = graphql_client.query(query)
-
-    assert result.data["user"] == "🍓"
-
-
-def test_hello(graphql_client):
-    query = """query {
         hello
     }
     """
@@ -18,14 +7,3 @@ def test_hello(graphql_client):
     result = graphql_client.query(query)
 
     assert result.data["hello"] == "🍓"
-
-
-def test_hello_field(graphql_client):
-    query = """query {
-        helloField
-    }
-    """
-
-    result = graphql_client.query(query)
-
-    assert result.data["helloField"] == "I'm a function resolver"
